Completed Tests and have developed a fully working GUI that meets 4 of my user stories so far.
Wrote code for saving and loading my program (the loading of my program still requires work and I can't get my tests to work because my case is very obscure and unique I can't seem to find a solution to getting the loading of my program to work.

Also have a fully working persistence package.

I now have a fully working save and load to file user story for my GUI, a piece of media (which is my introduction jingle), and am working CreateProfile and Adding Masses to a list.