Completed Tests and have developed a fully working GUI that meets every single one of my user stories.
Wrote code for saving and loading my program and fully implemented that in my GUI and officially have a fully working persistence package in console and the GUI.

Implemented a piece of media (which is my introduction jingle), and have a finished project with my required functionalities coded with their respective GUIs.

Made my user class robust as well.